Repository filtering, leaving the front only
