--- backend/src/app/app.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-@Module({
-  imports: [],
-  controllers: [],
-  providers: [],
-})
-export class AppModule {}
