+++ src/components/App.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function App() {
+  return <div className='App'></div>;
+}
+
+export default App;
